+++ backend/cartcraft/src/main/java/com/cart/cartcraft/request/AddProductRequest.java
@@ -4,13 +4,15 @@ import com.cart.cartcraft.model.Category;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class AddProductRequest {
     private Long id;
     private String name;
     private String brand;
-    private Double price;
-    private Integer quantity;
+    private BigDecimal price;
+    private int quantity;
     private String description;
 
     private Category category;
