+++ backend/cartcraft/src/main/java/com/cart/cartcraft/request/ProductUpdateRequest.java
@@ -3,13 +3,15 @@ package com.cart.cartcraft.request;
 import com.cart.cartcraft.model.Category;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductUpdateRequest {
     private Long id;
     private String name;
     private String brand;
-    private Double price;
-    private Integer quantity;
+    private BigDecimal price;
+    private int quantity;
     private String description;
 
     private Category category;
